move molecular_dipole to dipole module
