+++ pennylane/qchem/__init__.py
@@ -18,7 +18,7 @@ from .basis_data import load_basisset
 from .basis_set import BasisFunction, atom_basis_data, mol_basis_data
 from .convert import import_operator, import_state
 from .convert_openfermion import from_openfermion
-from .dipole import dipole_integrals, dipole_moment, fermionic_dipole
+from .dipole import dipole_integrals, dipole_moment, fermionic_dipole, molecular_dipole
 from .factorization import basis_rotation, factorize
 from .givens_decomposition import givens_decomposition
 from .hamiltonian import (
@@ -59,7 +59,6 @@ from .observable_hf import fermionic_observable, qubit_observable
 from .openfermion_pyscf import (
     decompose,
     dipole_of,
-    molecular_dipole,
     meanfield,
     observable,
     one_particle,
